requirements.txt
django-extra-views>=0.6.5
django-braces>=1.3.0
djangorestframework>=3.3.3
django-floppyforms>=1.6.2
django-filter>=0.13.0
django-crispy-forms>=1.3.2
django-debug-toolbar>=0.9.4
future>=0.15.2
pytz==2016.4
